chore: fmt

+++ testdata/zk/MockedModifier.t.sol
@@ -7,11 +7,7 @@ import "../cheats/Vm.sol";
 
 contract MockInner {
     // this covers an edge case that mainfests when returning >=5 items
-    function mockedMethod()
-        external
-        pure
-        returns (uint256, uint256, uint256, uint256, uint256)
-    {
+    function mockedMethod() external pure returns (uint256, uint256, uint256, uint256, uint256) {
         // We fail if this function isn't mocked
         assert(false);
         return (0, 0, 0, 0, 0);
@@ -27,7 +23,7 @@ contract Echoer {
 
     modifier needsMocking(uint256 n) {
         //we just check that we actually mock the value (to avoid optimization)
-        (, , uint r, , ) = mockInner.mockedMethod();
+        (,, uint256 r,,) = mockInner.mockedMethod();
         assert(r == n);
         _;
     }
@@ -40,16 +36,12 @@ contract Echoer {
         return n;
     }
 
-    function echo(
-        uint256[] memory n
-    ) external view needsMocking(42) returns (uint256[] memory) {
+    function echo(uint256[] memory n) external view needsMocking(42) returns (uint256[] memory) {
         assert(n.length == 1);
         return n;
     }
 
-    function echo(
-        Foo memory n
-    ) external view needsMocking(42) returns (Foo memory) {
+    function echo(Foo memory n) external view needsMocking(42) returns (Foo memory) {
         return n;
     }
 }
@@ -69,9 +61,7 @@ contract MockedModifierTest is DSTest {
         uint256 n = 10;
 
         vm.mockCall(
-            address(mockInner),
-            abi.encodeWithSelector(MockInner.mockedMethod.selector),
-            abi.encode(0, 0, 42, 0, 0, 0)
+            address(mockInner), abi.encodeWithSelector(MockInner.mockedMethod.selector), abi.encode(0, 0, 42, 0, 0, 0)
         );
 
         assertEq(n, target.echo(n));
@@ -82,11 +72,9 @@ contract MockedModifierTest is DSTest {
         n[0] = 10;
 
         vm.mockCall(
-            address(mockInner),
-            abi.encodeWithSelector(MockInner.mockedMethod.selector),
-            abi.encode(0, 0, 42, 0, 0, 0)
+            address(mockInner), abi.encodeWithSelector(MockInner.mockedMethod.selector), abi.encode(0, 0, 42, 0, 0, 0)
         );
-        
+
         assertEq(n[0], target.echo(n)[0]);
     }
 
@@ -94,11 +82,9 @@ contract MockedModifierTest is DSTest {
         Echoer.Foo memory n = Echoer.Foo({foo: 10});
 
         vm.mockCall(
-            address(mockInner),
-            abi.encodeWithSelector(MockInner.mockedMethod.selector),
-            abi.encode(0, 0, 42, 0, 0, 0)
+            address(mockInner), abi.encodeWithSelector(MockInner.mockedMethod.selector), abi.encode(0, 0, 42, 0, 0, 0)
         );
-        
+
         assertEq(n.foo, target.echo(n).foo);
     }
 }
